Change Datafile metadata from title to label (#50)
Fix bug. The Datafile metadata used the attribute `title` so far,
but this was not valid metadata attribute. After some research
and asking Phil Durbin from IQSS, the proper attribute seems to
be `label`, but there was no definitive answer. The fix was tested
locally on a Docker instance. After the upload of a testfile,
the showed filename on the frontend was overwritten with the
value coming from the `label` attribute.

Steps:

- [x] Update Datafile Tests

- [x] Update Datafile JSON Schema

- [x] Update Datafile Full JSON file

- [x] Update Datafile model

- [x] Tests: upload Datafile

Resolves: #50