Update Docs Occurrences of requests.Response
Replace all occurrences of requests.Reponse with httpx.Response.

Also do some minor documentation updates, including some formatting and removal
of example texts.
Include :special-members: for the API interface, as the auth-parameter will be
documented in the __init__ method, and the docs of the context manager for
async were not published yet.
Additionally, remove the intersphinx reference to requests. Unfortunately,
httpx does not have an inventory file (see also [1])

  [1]: https://github.com/encode/httpx/discussions/3091

Closes #198.