Contrib guide: fix missing titles and include
While building the docs with tox -e docs, ../CONTRIBUTING.rst was missing.
After fixing this, there was no link to the Contributor Guide due to a
missing title. So I added the title to contributing/contributing.rst and
included CONTRIBUTING.rst from line 3 to prevent the double title.
Finally, the language had to be set to non-None to avoid another WARNING.